+++ stdnet/orm/models.py
@@ -5,11 +5,13 @@ from stdnet.utils import zip, UnicodeMixin
 from stdnet import dispatch
 
 from .base import StdNetType
+from .globals import get_model_from_hash 
 from .signals import *
 
 
 __all__ = ['StdModel',
            'StdNetType',
+           'from_uuid',
            'model_to_dict']
 
 
@@ -86,6 +88,16 @@ otherwise a :class:`stdnet.exceptions.ModelNotRegistered` exception will raise.'
         if not self.id:
             raise self.DoesNotExist('Object not saved. Cannot obtain universally unique id')
         return '{0}.{1}'.format(self._meta.hash,self.id)
+    
+    @classmethod
+    def from_uuid(self, uuid):
+        elems = uuid.split('.')
+        if len(elems) == 2:
+            model = get_model_from_hash(elems[0])
+            if not model:
+                raise cls.DoesNotExist('model id "{0}" not available'.format(elems[0]))
+            return model.objects.get(id = elems[1])
+        raise cls.DoesNotExist('uuid "{0}" not recognized'.format(uuid))
         
     def __eq__(self, other):
         if other.__class__ == self.__class__:
@@ -178,6 +190,7 @@ will enumerate the number of object to delete. without deleting them.'''
         self.afterload()
     
 
+from_uuid = StdModel.from_uuid
 
 def model_to_dict(instance, fields = None, exclude = None):
     if isinstance(instance,StdModel):